+++ map/src/tiles/mvt/mvt_parser.rs
@@ -1,10 +1,9 @@
-use crate::MAX_ZOOM_LEVEL;
 use crate::tiles::mvt::mvt_scheme_parser::MvtSchemeParser;
+use crate::MAX_ZOOM_LEVEL;
 use fast_mvt::proto::GeomType;
-use fast_mvt::{MvtGeometry, MvtReaderRef, MvtResult, MvtValue};
-use geo::{CoordsIter, MapCoords};
-use geo_types::{Geometry, LineString, Polygon, coord};
-use log::error;
+use fast_mvt::{MvtReaderRef, MvtResult};
+use geo::MapCoords;
+use geo_types::{Coord, Geometry, LineString, Polygon};
 use osm::map::{MapGeomObject, MapGeometry};
 use osm::tiles::TileKey;
 
@@ -28,13 +27,22 @@ impl MvtParser {
         }
     }
 
+    fn get_all_polygons(geometry: Geometry<i32>) -> Vec<Polygon<i32>> {
+        match geometry {
+            Geometry::Polygon(polygon) => {
+                vec![polygon]
+            }
+            Geometry::MultiPolygon(multi_polygon) => multi_polygon.into_iter().collect(),
+            _ => Vec::new(),
+        }
+    }
+
     pub fn read_mvt_tile(
         &self,
         bytes: &[u8],
         tile_key: &TileKey,
     ) -> MvtResult<Vec<(MapGeomObject, MapGeometry<f32>)>> {
         let reader = MvtReaderRef::new(bytes)?;
-        let koef = 512.0f32 / 4096.0;
         let mut result = self.schema_parser.parse(reader.layers(), |feature| {
             let mut res = vec![];
             let geom_type = feature.geom_type();
@@ -44,34 +52,14 @@ impl MvtParser {
                 && geometry.is_ok()
             {
                 let geometry = geometry.unwrap();
+                // TODO Add points later
                 if geom_type == GeomType::LINESTRING {
                     for line in Self::get_all_lines(geometry) {
-                        let ls: LineString<f32> = line
-                            .coords_iter()
-                            .map(|coord| {
-                                coord! { x: coord.x as f32 * koef, y: coord.y as f32 * koef}
-                            })
-                            .collect();
-
-                        res.push(MapGeometry::Line(ls));
+                        res.push(MapGeometry::Line(line));
                     }
                 } else if geom_type == GeomType::POLYGON {
-                    match geometry {
-                        MvtGeometry::Polygon(poly) => {
-                            let ls: Polygon<f32> = poly.map_coords(|coord| {
-                                coord! { x: coord.x as f32 * koef, y: coord.y as f32 * koef}
-                            });
-                            res.push(MapGeometry::Poly(ls));
-                        }
-                        MvtGeometry::MultiPolygon(polis) => {
-                            for poly in polis {
-                                let ls: Polygon<f32> = poly.map_coords(|coord| {
-                                    coord! { x: coord.x as f32 * koef, y: coord.y as f32 * koef}
-                                });
-                                res.push(MapGeometry::Poly(ls));
-                            }
-                        }
-                        _ => {}
+                    for polygon in Self::get_all_polygons(geometry) {
+                        res.push(MapGeometry::Poly(polygon));
                     }
                 }
             }
@@ -91,50 +79,23 @@ impl MvtParser {
     }
 
     fn convert_and_restore_data(
-        geometry: &MapGeometry<f32>,
+        geometry: &MapGeometry<i32>,
         tile_key: &TileKey,
     ) -> MapGeometry<f32> {
         // FIXME Zoom level handling
         let factor = 2.0f32.powf((MAX_ZOOM_LEVEL - tile_key.zoom_level) as f32);
+        let koef = 512.0f32 / 4096.0;
+        let total_multiplier = factor * koef;
+        let convert_coord = |c: &Coord<i32>| -> Coord<f32> {
+            Coord {
+                x: (c.x as f32) * total_multiplier,
+                y: (c.y as f32) * total_multiplier,
+            }
+        };
         match geometry {
-            MapGeometry::Line(line) => MapGeometry::Line(line.map_coords(|coord| coord * factor)),
-            MapGeometry::Poly(poly) => MapGeometry::Poly(poly.map_coords(|coord| coord * factor)),
-            MapGeometry::Coord(coord) => MapGeometry::Coord(*coord * factor),
-        }
-    }
-}
-
-pub(crate) struct LocalMvtValue(pub MvtValue);
-
-impl LocalMvtValue {
-    fn unexpected_type<T>(&self, expected: &str) -> Option<T> {
-        error!("Unexpected {} MvtValueRef: {:?}", expected, self.0);
-        None
-    }
-}
-impl From<LocalMvtValue> for Option<i64> {
-    fn from(value: LocalMvtValue) -> Self {
-        match value.0 {
-            MvtValue::SInt(value) => Some(value),
-            _ => value.unexpected_type("i64"),
-        }
-    }
-}
-
-impl From<LocalMvtValue> for Option<String> {
-    fn from(value: LocalMvtValue) -> Self {
-        match value.0 {
-            MvtValue::String(value) => Some(value),
-            _ => value.unexpected_type("String"),
-        }
-    }
-}
-
-impl From<LocalMvtValue> for Option<bool> {
-    fn from(value: LocalMvtValue) -> Self {
-        match value.0 {
-            MvtValue::Bool(value) => Some(value),
-            _ => value.unexpected_type("bool"),
+            MapGeometry::Line(line) => MapGeometry::Line(line.map_coords(|c| convert_coord(&c))),
+            MapGeometry::Poly(poly) => MapGeometry::Poly(poly.map_coords(|c| convert_coord(&c))),
+            MapGeometry::Coord(coord) => MapGeometry::Coord(convert_coord(coord)),
         }
     }
 }
